refactor(nap-serv): add barrel exports for cross-boundary service APIs
Cross-module consumers (ap, ar, projects) now import from barrel
index.js files instead of reaching into sibling module internals.
Defines a stable public API surface for core/services (allocateNumber)
and accounting/services (postAPInvoice, postAPPayment, etc.).

Intra-module consumers keep their direct imports unchanged.
