+++ apps/nap-serv/src/system/core/services/index.js
@@ -1,0 +1,11 @@
+/**
+ * @file Core services barrel — stable cross-boundary API
+ * @module core/services
+ *
+ * Re-exports the public API for consumers outside the core module.
+ * Intra-module consumers may import directly from individual files.
+ *
+ * Copyright (c) 2025 NapSoft LLC. All rights reserved.
+ */
+
+export { allocateNumber } from './numberingService.js';
